requirement.txt
sqlalchemy
beautifultable
python-decouple
aiohttp